chore: rename git_push_stack to mergify_cli
Mergify-cli stack change-Id: I97a1959ad3bbc102dcffd93b3a2271dc8c097f7d:
